+++ app/src/renderer/components/staking/PageValidator.vue
@@ -34,26 +34,68 @@ tm-page(:title='validatorTitle(this.validator)')
       tm-list-item(dt="Commission Maximum" :dd="pretty(validator.commission_max) + ' %'")
       tm-list-item(dt="Commission Change-Rate" :dd="pretty(validator.commission_change_rate) + ' %'")
       tm-list-item(dt="Commission Change Today" :dd="pretty(validator.commission_change_today) + ' %'")
+
+    modal-stake(
+      v-if="showModalStake"
+      v-on:submitDelegation="submitDelegation"
+      :showModalStake.sync="showModalStake"
+      :fromOptions="[{ key: `My Wallet - ${this.wallet.address}`, value: 0 }]"
+      :maximum="availableAtoms()"
+      :to="validator.owner"
+    )
+
+    tm-modal(:close="closeCannotStake" icon="warning" v-if="showCannotStake")
+      div(slot='title') Cannot Stake
+      p You have no {{ bondingDenom }}s to stake.
+      div(slot='footer')
+        tmBtn(@click.native="closeCannotStake()" value="OK")
+
+    tm-btn(
+      @click.native="onStake()"
+      color="primary"
+      id="Stake"
+      value="Stake"
+    )
 </template>
 
 <script>
 import { mapGetters } from "vuex"
-import { TmListItem, TmPage, TmPart, TmToolBar } from "@tendermint/ui"
+import { TmBtn, TmListItem, TmPage, TmPart, TmToolBar } from "@tendermint/ui"
 import { TmDataError } from "common/TmDataError"
+import { calculateTokens } from "scripts/common"
+import ModalStake from "staking/ModalStake"
 import numeral from "numeral"
 import AnchorCopy from "common/AnchorCopy"
+import TmModal from "common/TmModal"
+
 export default {
   name: "page-validator",
   components: {
     AnchorCopy,
+    ModalStake,
+    TmBtn,
     TmListItem,
+    TmModal,
     TmPage,
     TmPart,
     TmToolBar,
     TmDataError
   },
+  data: () => ({
+    showCannotStake: false,
+    showModalStake: false
+  }),
   computed: {
-    ...mapGetters(["delegates", "config", "keybase"]),
+    ...mapGetters([
+      `bondingDenom`,
+      "delegates",
+      `delegation`,
+      "config",
+      "keybase",
+      `oldBondedAtoms`,
+      `totalAtoms`,
+      `wallet`
+    ]),
     validator() {
       let validator = this.delegates.delegates.find(
         v => this.$route.params.validator === v.owner
@@ -68,6 +110,61 @@ export default {
     }
   },
   methods: {
+    availableAtoms() {
+      return this.totalAtoms - this.oldBondedAtoms
+    },
+    closeCannotStake() {
+      this.showCannotStake = false
+    },
+    onStake() {
+      if (this.availableAtoms() > 0) {
+        this.showModalStake = true
+      } else {
+        this.showCannotStake = true
+      }
+    },
+    async submitDelegation({ amount }) {
+      const candidateId = this.validator.owner
+
+      const currentlyDelegated = calculateTokens(
+        this.validator,
+        this.delegation.committedDelegates[candidateId] || 0
+      )
+
+      const delegation = [
+        {
+          atoms: currentlyDelegated.plus(amount),
+          delegate: this.validator
+        }
+      ]
+
+      try {
+        await this.$store.dispatch("submitDelegation", delegation)
+
+        this.$store.commit("notify", {
+          title: "Successful Staking!",
+          body: `You have successfully staked your ${this.bondingDenom}s.`
+        })
+      } catch (exception) {
+        const { message } = exception
+        console.log(exception.stack)
+        let errData = message.split("\n")[5]
+
+        if (errData) {
+          let parsedErr = errData.split('"')[1]
+
+          this.$store.commit("notifyError", {
+            title: `Error While Staking ${this.bondingDenom}s`,
+            body: parsedErr[0].toUpperCase() + parsedErr.slice(1)
+          })
+        } else {
+          this.$store.commit("notifyError", {
+            title: `Error While Staking ${this.bondingDenom}s`,
+            body: message
+          })
+        }
+      }
+    },
     validatorTitle(validator) {
       if (!validator) return "Validator Not Found"
       let title