+++ test/unit/specs/components/staking/PageValidator.spec.js
@@ -1,5 +1,55 @@
+import BigNumber from "bignumber.js"
+import ModalStake from "staking/ModalStake"
 import setup from "../../../helpers/vuex-setup"
 import PageValidator from "renderer/components/staking/PageValidator"
+import { mount } from "@vue/test-utils"
+
+// Create a getters object from an object of simple values.
+const mockGetters = values =>
+  Object.assign(
+    ...Object.entries(values).map(([key, value]) => ({ [key]: () => value }))
+  )
+
+const delegate = {
+  owner: "1a2b3c",
+  pub_key: {
+    type: "AC26791624DE60",
+    data: "dlN5SLqeT3LT9WsUK5iuVq1eLQV2Q1JQAuyN0VwSWK0="
+  },
+  tokens: "19",
+  delegator_shares: "19",
+  description: {
+    description: "Herr Schmidt",
+    moniker: "herr_schmidt_revoked",
+    country: "DE"
+  },
+  revoked: true,
+  status: 2,
+  bond_height: "0",
+  bond_intra_tx_counter: 6,
+  proposer_reward_pool: null,
+  commission: "0",
+  commission_max: "0",
+  commission_change_rate: "0",
+  commission_change_today: "0",
+  prev_bonded_shares: "0"
+}
+
+const getterValues = {
+  bondingDenom: `atom`,
+  config: { desktop: false },
+  delegates: {
+    delegates: [delegate]
+  },
+  delegation: {
+    committedDelegates: { "1a2b3c": 0 },
+    unbondingDelegations: {}
+  },
+  keybase: `keybase`,
+  oldBondedAtoms: 50,
+  totalAtoms: 100,
+  wallet: { address: `cosmosaccaddr15ky9du8a2wlstz6fpx3p4mqpjyrm5ctpesxxn9` }
+}
 
 describe("PageValidator", () => {
   let wrapper
@@ -10,37 +60,7 @@ describe("PageValidator", () => {
       doBefore: ({ router }) => {
         router.push("/staking/validators/1a2b3c")
       },
-      getters: {
-        config: () => ({ desktop: false }),
-        delegates: () => ({
-          delegates: [
-            {
-              owner: "1a2b3c",
-              pub_key: {
-                type: "AC26791624DE60",
-                data: "dlN5SLqeT3LT9WsUK5iuVq1eLQV2Q1JQAuyN0VwSWK0="
-              },
-              tokens: "19",
-              delegator_shares: "19",
-              description: {
-                description: "Herr Schmidt",
-                moniker: "herr_schmidt_revoked",
-                country: "DE"
-              },
-              revoked: true,
-              status: 2,
-              bond_height: "0",
-              bond_intra_tx_counter: 6,
-              proposer_reward_pool: null,
-              commission: "0",
-              commission_max: "0",
-              commission_change_rate: "0",
-              commission_change_today: "0",
-              prev_bonded_shares: "0"
-            }
-          ]
-        })
-      }
+      getters: mockGetters(getterValues)
     })
     wrapper = instance.wrapper
   })
@@ -121,5 +141,159 @@ describe("PageValidator", () => {
     wrapper = instance.wrapper
 
     expect(wrapper.vm.$el).toMatchSnapshot()
+  })
+})
+
+describe(`onStake`, () => {
+  describe(`make sure we have enough atoms to stake`, () => {
+    it(`is enough`, () => {
+      const $store = {
+        commit: jest.fn(),
+        dispatch: jest.fn(),
+        getters: getterValues
+      }
+
+      const wrapper = mount(PageValidator, {
+        mocks: {
+          $route: { params: { validator: `1a2b3c` } },
+          $store
+        }
+      })
+
+      wrapper.find(`#Stake`).trigger(`click`)
+      expect(wrapper.contains(ModalStake)).toEqual(true)
+    })
+
+    it(`is not enough`, () => {
+      const $store = {
+        commit: jest.fn(),
+        dispatch: jest.fn(),
+        getters: Object.assign({}, getterValues, { oldBondedAtoms: 100 })
+      }
+
+      const wrapper = mount(PageValidator, {
+        mocks: {
+          $route: { params: { validator: `1a2b3c` } },
+          $store
+        }
+      })
+
+      wrapper.find(`#Stake`).trigger(`click`)
+
+      expect(wrapper.text().includes(`You have no atoms to stake.`)).toEqual(
+        true
+      )
+
+      wrapper.find(`button`).trigger(`click`)
+
+      expect(wrapper.text().includes(`You have no atoms to stake.`)).toEqual(
+        false
+      )
+    })
+  })
+
+  describe(`submitDelegation`, () => {
+    it(`success`, async () => {
+      const $store = {
+        commit: jest.fn(),
+        dispatch: jest.fn(),
+        getters: getterValues
+      }
+
+      const {
+        vm: { submitDelegation }
+      } = mount(PageValidator, {
+        mocks: {
+          $route: { params: { validator: `1a2b3c` } },
+          $store
+        }
+      })
+
+      await submitDelegation({ amount: 10 })
+
+      expect($store.dispatch.mock.calls).toEqual([
+        [`submitDelegation`, [{ atoms: BigNumber(10), delegate }]]
+      ])
+
+      expect($store.commit.mock.calls).toEqual([
+        [
+          `notify`,
+          {
+            body: `You have successfully staked your atoms.`,
+            title: `Successful Staking!`
+          }
+        ]
+      ])
+    })
+
+    it(`error`, async () => {
+      const $store = {
+        commit: jest.fn(),
+        dispatch: jest.fn(() => {
+          throw new Error(`message`)
+        }),
+        getters: getterValues
+      }
+
+      const {
+        vm: { submitDelegation }
+      } = mount(PageValidator, {
+        mocks: {
+          $route: { params: { validator: `1a2b3c` } },
+          $store
+        }
+      })
+
+      await submitDelegation({ amount: 10 })
+
+      expect($store.dispatch.mock.calls).toEqual([
+        [`submitDelegation`, [{ atoms: BigNumber(10), delegate }]]
+      ])
+
+      expect($store.commit.mock.calls).toEqual([
+        [
+          `notifyError`,
+          {
+            body: `message`,
+            title: `Error While Staking atoms`
+          }
+        ]
+      ])
+    })
+
+    it(`error with data`, async () => {
+      const $store = {
+        commit: jest.fn(),
+        dispatch: jest.fn(() => {
+          throw new Error(`one\ntwo\nthree\nfour\nfive\nsix"seven`)
+        }),
+        getters: getterValues
+      }
+
+      const {
+        vm: { submitDelegation }
+      } = mount(PageValidator, {
+        mocks: {
+          $route: { params: { validator: `1a2b3c` } },
+          $store
+        }
+      })
+
+      await submitDelegation({ amount: 10 })
+
+      expect($store.dispatch.mock.calls).toEqual([
+        [`submitDelegation`, [{ atoms: BigNumber(10), delegate }]]
+      ])
+
+      expect($store.commit.mock.calls).toEqual([
+        [
+          `notifyError`,
+          {
+            body: `Seven`,
+            title: `Error While Staking atoms`
+          }
+        ]
+      ])
+    })
   })
 })