+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Refactored Addressbook. @NodeGuy
 
+### Changed
+
+* Created new modal window for staking to a validator. @NodeGuy
+
 ### Fixed
 
 * Do not try to publish a release on every push to `develop`! @NodeGuy